## Changes

- Removed ConstraintLayout dependency
- Simplified layout hierarchy with Box, Column, and Row


## Result Comparison (Before & After)

### Macrobenchmark Results (Before)

| Metric                    | Min   | Median | Max   | P50   | P90   | P95   | P99   |
|---------------------------|-------|--------|-------|-------|-------|-------|-------|
| frameCount                | 156.0 | 165.0  | 233.0 |   –   |   –   |   –   |   –   |
| timeToInitialDisplayMs    | 540.4 | 558.2  | 571.5 |   –   |   –   |   –   |   –   |
| frameDurationCpuMs        |   –   |   –    |   –   | 6.7   | 23.0  | 31.4  | 74.5  |
| frameOverrunMs            |   –   |   –    |   –   | -3.5  | 14.3  | 30.9  | 307.2 |


---

### Macrobenchmark Results (After)

| Metric                    | Min   | Median | Max   | P50   | P90   | P95   | P99   |
|---------------------------|-------|--------|-------|-------|-------|-------|-------|
| frameCount                | 135.0 | 157.0  | 218.0 |   –   |   –   |   –   |   –   |
| timeToInitialDisplayMs    | 523.3 | 556.7  | 710.4 |   –   |   –   |   –   |   –   |
| frameDurationCpuMs        |   –   |   –    |   –   | 6.9   | 22.1  | 27.2  | 67.1  |
| frameOverrunMs            |   –   |   –    |   –   | -3.0  | 13.8  | 33.6  | 283.9 |


### Perfetto Frame Timing Comparison

| Metric                | Before         | After          |
|-----------------------|---------------|----------------|
| First Frame Duration  | 317 ms 508 μs | 304 ms 8 μs    |
| Measure Phase Time    | 129 ms 88 μs  | 112 ms 744 μs  |

- [Before Trace (Google Drive)](https://drive.google.com/file/d/1Bpau8eYZ_t9CuEbFAlBoIKl46rm4imXo/view?usp=sharing)
- [After Trace (Google Drive)](https://drive.google.com/file/d/1ijZo13xuy--aILhohFmVXlxLX1I_PVWa/view?usp=sharing)


## Key Findings

- At first, I thought the layout hierachy within ConstraintLayout in Composable caused impact frame rendering delay, but the difference turns out to be slight.
- I found that Recomposition is the primary cause of the delays.